# server/ui/api/v1/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "OpenAPI 2.0 documentation for [MessagePit](https://github.com/coreydaley/messagepit).",
    "title": "MessagePit API",
    "contact": {
      "name": "GitHub",
      "url": "https://github.com/coreydaley/messagepit"
    },
    "license": {
      "name": "MIT license",
      "url": "https://github.com/coreydaley/messagepit/blob/main/LICENSE"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/message/{ID}": {
      "get": {
        "description": "Returns the summary of a message, marking the message as read.\n\nThe ID can be set to `latest` to return the latest message.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "message"
        ],
        "summary": "Get message summary",
        "operationId": "GetMessageParams",
        "parameters": [
          {
            "type": "string",
            "description": "Message database ID or \"latest\"",
            "name": "ID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Message",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundResponse"
          }
        }
      }
    },
    "/api/v1/message/{ID}/headers": {
      "get": {
        "description": "Returns the message headers as an array. Note that header keys are returned alphabetically.\n\nThe ID can be set to `latest` to return the latest message headers.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "message"
        ],
        "summary": "Get message headers",
        "operationId": "GetHeadersParams",
        "parameters": [
          {
            "type": "string",
            "description": "Message database ID or \"latest\"",
            "name": "ID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "MessageHeadersResponse",
            "schema": {
              "$ref": "#/definitions/MessageHeadersResponse"
            }
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundResponse"
          }
        }
      }
    },
...
    "/api/v1/message/{ID}/part/{PartID}": {
      "get": {
        "description": "This will return the attachment part using the appropriate Content-Type.\n\nThe ID can be set to `latest` to reference the latest message.",
        "produces": [
          "application/*",
          "image/*",
          "text/*"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "message"
        ],
        "summary": "Get message attachment",
        "operationId": "AttachmentParams",
        "parameters": [
          {
            "type": "string",
            "description": "Message database ID or \"latest\"",
            "name": "ID",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "Attachment part ID",
            "name": "PartID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/BinaryResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundResponse"
          }
        }
      }
    },
...
    "/api/v1/message/{ID}/raw": {
      "get": {
        "description": "Returns the full email source as plain text.\n\nThe ID can be set to `latest` to return the latest message source.",
        "produces": [
          "text/plain"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "message"
        ],
        "summary": "Get message source",
        "operationId": "DownloadRawParams",
        "parameters": [
          {
            "type": "string",
            "description": "Message database ID or \"latest\"",
            "name": "ID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/TextResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundResponse"
          }
        }
      }
    },
...
    "/api/v1/messages": {
      "get": {
        "description": "Returns messages from the mailbox ordered from newest to oldest.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "messages"
        ],
        "summary": "List messages",
        "operationId": "GetMessagesParams",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "default": 0,
            "x-go-name": "Start",
            "description": "Pagination offset",
            "name": "start",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int64",
            "default": 50,
            "x-go-name": "Limit",
            "description": "Limit number of results",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/MessagesSummaryResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      },
      "put": {
        "description": "You can optionally provide an array of IDs or a search string.\nIf neither IDs nor search is provided then all mailbox messages are updated.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "text/plain"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "messages"
        ],
        "summary": "Set read status",
        "operationId": "SetReadStatusParams",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "IDs": {
                  "description": "Optional array of message database IDs",
                  "type": "array",
                  "default": [],
                  "items": {
                    "type": "string"
                  },
                  "example": [
                    "4oRBnPtCXgAqZniRhzLNmS",
                    "hXayS6wnCgNnt6aFTvmOF6"
                  ]
                },
                "Read": {
                  "description": "Read status",
                  "type": "boolean",
                  "default": false,
                  "example": true
                },
                "Search": {
                  "description": "Optional messages matching a search",
                  "type": "string",
                  "example": "tag:backups"
                }
              }
            }
          },
          {
            "type": "string",
            "x-go-name": "TZ",
            "description": "Optional [timezone identifier](https://en.wikipedia.org/wiki/List_of_tz_database_time_zones) used only for `before:` \u0026 `after:` searches (eg: \"Pacific/Auckland\").",
            "name": "tz",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/OKResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      },
      "delete": {
        "description": "Delete individual or all messages. If no IDs are provided then all messages are deleted.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "text/plain"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "messages"
        ],
        "summary": "Delete messages",
        "operationId": "DeleteMessagesParams",
        "parameters": [
          {
            "description": "Delete request",
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "IDs": {
                  "description": "Array of message database IDs",
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "example": [
                    "4oRBnPtCXgAqZniRhzLNmS",
                    "hXayS6wnCgNnt6aFTvmOF6"
                  ]
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/OKResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/search": {
      "get": {
        "description": "Returns messages matching [a search](https://mailpit.axllent.org/docs/usage/search-filters/), sorted by received date (descending).",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "messages"
        ],
        "summary": "Search messages",
        "operationId": "SearchParams",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Query",
            "description": "Search query",
            "name": "query",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "default": "0",
            "x-go-name": "Start",
            "description": "Pagination offset",
            "name": "start",
            "in": "query"
          },
          {
            "type": "string",
            "default": "50",
            "x-go-name": "Limit",
            "description": "Limit results",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "TZ",
            "description": "Optional [timezone identifier](https://en.wikipedia.org/wiki/List_of_tz_database_time_zones) used only for `before:` \u0026 `after:` searches (eg: \"Pacific/Auckland\").",
            "name": "tz",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/MessagesSummaryResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      },
      "delete": {
        "description": "Delete all messages matching [a search](https://mailpit.axllent.org/docs/usage/search-filters/).",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "messages"
        ],
        "summary": "Delete messages by search",
        "operationId": "DeleteSearchParams",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Query",
            "description": "Search query",
            "name": "query",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "TZ",
            "description": "[Timezone identifier](https://en.wikipedia.org/wiki/List_of_tz_database_time_zones) used only for `before:` \u0026 `after:` searches (eg: \"Pacific/Auckland\").",
            "name": "tz",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/OKResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Address": {
      "description": "An address such as \"Barry Gibbs \u003cbg@example.com\u003e\" is represented\nas Address{Name: \"Barry Gibbs\", Address: \"bg@example.com\"}.",
      "type": "object",
      "title": "Address represents a single mail address.",
      "properties": {
        "Address": {
          "type": "string"
        },
        "Name": {
          "type": "string"
        }
      },
      "x-go-package": "net/mail"
    },
...
    "Attachment": {
      "description": "Attachment struct for inline images and attachments",
      "type": "object",
      "properties": {
        "Checksums": {
          "description": "File checksums",
          "type": "object",
          "properties": {
            "MD5": {
              "description": "MD5 checksum hash of file",
              "type": "string"
            },
            "SHA1": {
              "description": "SHA1 checksum hash of file",
              "type": "string"
            },
            "SHA256": {
              "description": "SHA256 checksum hash of file",
              "type": "string"
            }
          }
        },
        "ContentID": {
          "description": "Content ID",
          "type": "string"
        },
        "ContentType": {
          "description": "Content type",
          "type": "string"
        },
        "FileName": {
          "description": "File name",
          "type": "string"
        },
        "PartID": {
          "description": "Attachment part ID",
          "type": "string"
        },
        "Size": {
          "description": "Size in bytes",
          "type": "integer",
          "format": "uint64"
        }
      },
      "x-go-package": "github.com/coreydaley/messagepit/internal/storage"
    },
...
    "ListUnsubscribe": {
      "description": "ListUnsubscribe contains a summary of List-Unsubscribe \u0026 List-Unsubscribe-Post headers\nincluding validation of the link structure",
      "type": "object",
      "properties": {
        "Errors": {
          "description": "Validation errors (if any)",
          "type": "string"
        },
        "Header": {
          "description": "List-Unsubscribe header value",
          "type": "string"
        },
        "HeaderPost": {
          "description": "List-Unsubscribe-Post value (if set)",
          "type": "string"
        },
        "Links": {
          "description": "Detected links, maximum one email and one HTTP(S) link",
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "x-go-package": "github.com/coreydaley/messagepit/internal/storage"
    },
    "Message": {
      "description": "Message data excluding physical attachments",
      "type": "object",
      "properties": {
        "Attachments": {
          "description": "Message attachments",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Attachment"
          }
        },
        "Bcc": {
          "description": "Bcc addresses",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Address"
          }
        },
        "Cc": {
          "description": "Cc addresses",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Address"
          }
        },
        "Date": {
          "description": "Message RFC3339Nano date \u0026 time (if set), else date \u0026 time received\n([extended RFC3339](https://tools.ietf.org/html/rfc3339#section-5.6) format with optional nano seconds)",
          "type": "string",
          "format": "date-time"
        },
        "From": {
          "$ref": "#/definitions/Address"
        },
        "HTML": {
          "description": "Message body HTML",
          "type": "string"
        },
        "ID": {
          "description": "Database ID",
          "type": "string"
        },
        "Inline": {
          "description": "Inline message attachments",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Attachment"
          }
        },
        "ListUnsubscribe": {
          "$ref": "#/definitions/ListUnsubscribe"
        },
        "MessageID": {
          "description": "Message ID",
          "type": "string"
        },
        "ReplyTo": {
          "description": "ReplyTo addresses",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Address"
          }
        },
        "ReturnPath": {
          "description": "Return-Path",
          "type": "string"
        },
        "Size": {
          "description": "Message size in bytes",
          "type": "integer",
          "format": "uint64"
        },
        "Subject": {
          "description": "Message subject",
          "type": "string"
        },
        "Tags": {
          "description": "Message tags",
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "Text": {
          "description": "Message body text",
          "type": "string"
        },
        "To": {
          "description": "To addresses",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Address"
          }
        },
        "Username": {
          "description": "Username used for authentication (if provided) with the SMTP or Send API",
          "type": "string"
        }
      },
      "x-go-package": "github.com/coreydaley/messagepit/internal/storage"
    },
    "MessageHeadersResponse": {
      "description": "Message headers",
      "type": "object",
      "additionalProperties": {
        "type": "array",
        "items": {
          "type": "string"
        }
      },
      "x-go-name": "messageHeadersResponse",
      "x-go-package": "github.com/coreydaley/messagepit/server/apiv1"
    },
    "MessageSummary": {
      "description": "MessageSummary struct for frontend messages",
      "type": "object",
      "properties": {
        "Attachments": {
          "description": "Whether the message has any attachments",
          "type": "integer",
          "format": "int64"
        },
        "Bcc": {
          "description": "Bcc addresses",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Address"
          }
        },
        "Cc": {
          "description": "Cc addresses",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Address"
          }
        },
        "Created": {
          "description": "Received RFC3339Nano date \u0026 time ([extended RFC3339](https://tools.ietf.org/html/rfc3339#section-5.6) format with optional nano seconds)",
          "type": "string",
          "format": "date-time"
        },
        "From": {
          "$ref": "#/definitions/Address"
        },
        "ID": {
          "description": "Database ID",
          "type": "string"
        },
        "MessageID": {
          "description": "Message ID",
          "type": "string"
        },
        "Read": {
          "description": "Read status",
          "type": "boolean"
        },
        "ReplyTo": {
          "description": "Reply-To address",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Address"
          }
        },
        "Size": {
          "description": "Message size in bytes (total)",
          "type": "integer",
          "format": "uint64"
        },
        "Snippet": {
          "description": "Message snippet includes up to 250 characters",
          "type": "string"
        },
        "Subject": {
          "description": "Email subject",
          "type": "string"
        },
        "Tags": {
          "description": "Message tags",
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "To": {
          "description": "To address",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Address"
          }
        },
        "Username": {
          "description": "Username used for authentication (if provided) with the SMTP or Send API",
          "type": "string"
        }
      },
      "x-go-package": "github.com/coreydaley/messagepit/internal/storage"
    },
    "MessagesSummary": {
      "description": "MessagesSummary is a summary of a list of messages",
      "type": "object",
      "properties": {
        "messages": {
          "description": "Messages summary\nin: body",
          "type": "array",
          "items": {
            "$ref": "#/definitions/MessageSummary"
          },
          "x-go-name": "Messages"
        },
        "messages_count": {
          "description": "Total number of messages matching current query",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "MessagesCount"
        },
        "messages_unread": {
          "description": "Total number of unread messages matching current query",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "MessagesUnreadCount"
        },
        "start": {
          "description": "Pagination offset",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Start"
        },
        "tags": {
          "description": "All current tags",
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Tags"
        },
        "total": {
          "description": "Total number of messages in mailbox",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "Total"
        },
        "unread": {
          "description": "Total number of unread messages in mailbox",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "Unread"
        }
      },
      "x-go-package": "github.com/coreydaley/messagepit/server/apiv1"
    },
...
  },
  "responses": {
...
    "BinaryResponse": {
      "description": "Binary data response which inherits the attachment's content type.",
      "schema": {
        "type": "string"
      }
    },
...
    "ErrorResponse": {
      "description": "Server error will return with a 400 status code\nwith the error message in the body",
      "schema": {
        "type": "string"
      }
    },
...
    "MessagesSummaryResponse": {
      "description": "Summary of messages",
      "schema": {
        "$ref": "#/definitions/MessagesSummary"
      }
    },
    "NotFoundResponse": {
      "description": "Not found error will return a 404 status code",
      "schema": {
        "type": "string"
      }
    },
    "OKResponse": {
      "description": "Plain text \"ok\" response",
      "schema": {
        "type": "string"
      }
    },
...
    "TextResponse": {
      "description": "Plain text response",
      "schema": {
        "type": "string"
      }
    },
...
  }
}